+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Minimum PHP version requirement has been raised from 7.0 to 7.4.
 - Upgraded Swiftmailer to version 6.3.
+- PHP's built-in `mail()` function can no longer be used. Anyone using the `mail` protocol will now use Sendmail instead.
 - Moved the errors directory from `application/views` into the Nova core.
 
 ### Fixed